transitNetwork backend: merge executor and algorithm classes
They have the same role of controlling the algorithm run, so they are
now one single class, extending the JobWrapper
